Add `eslint-plugin-testing-library` to React configuration (#92)
This PR adds eslint-plugin-testing-library to the react configuration and requires it as a peer dependency. The plugin has been added to an override that targets tests only, as described in the project's readme.

Applied the regex recommended by the testing-library eslint rule's docs (which is based on Jest's defaults) to the shared config. To do this, I've split its override in two: a general override covering JavaScript and TypeScript test suites and another that is specific to TypeScript.